Sqlite/requirements.txt
appdirs==1.4.4
astroid==2.4.1
attrs==19.3.0
black==19.10b0
cblack==0.9
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.8.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.2
PyMySQL==0.9.3
python-telegram-bot==12.7
regex==2020.5.14
six==1.15.0
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
wrapt==1.12.1
